+++ kubernetes-itests/src/test/java/io/fabric8/kubernetes/IngressIT.java
@@ -17,6 +17,9 @@
 package io.fabric8.kubernetes;
 
 import io.fabric8.commons.ClusterEntity;
+import io.fabric8.commons.ConditionalIgnore;
+import io.fabric8.commons.ConditionalIgnoreRule;
+import io.fabric8.commons.K8sVersionLessThan_1_14;
 import io.fabric8.kubernetes.api.model.networking.v1beta1.Ingress;
 import io.fabric8.kubernetes.api.model.networking.v1beta1.IngressList;
 import io.fabric8.kubernetes.client.KubernetesClient;
@@ -26,6 +29,7 @@ import org.arquillian.cube.requirement.ArquillianConditionalRunner;
 import org.jboss.arquillian.test.api.ArquillianResource;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
+import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
@@ -37,6 +41,9 @@ import static org.junit.Assert.assertNotNull;
 @RunWith(ArquillianConditionalRunner.class)
 @RequiresKubernetes
 public class IngressIT {
+  @Rule
+  public ConditionalIgnoreRule rule = new ConditionalIgnoreRule();
+
   @ArquillianResource
   KubernetesClient client;
 
@@ -47,10 +54,15 @@ public class IngressIT {
 
   @BeforeClass
   public static void init() {
+    // Ignore if Kubernetes version < 1.14
+    if (new K8sVersionLessThan_1_14().isSatisfied()) {
+      return;
+    }
     ClusterEntity.apply(IngressIT.class.getResourceAsStream("/ingress-it.yml"));
   }
 
   @Test
+  @ConditionalIgnore(condition = K8sVersionLessThan_1_14.class)
   public void load() {
     Ingress aIngress = client.network().v1beta1().ingresses().inNamespace(session.getNamespace()).load(getClass().getResourceAsStream("/test-ingress.yml")).get();
     assertThat(aIngress).isNotNull();
@@ -58,12 +70,14 @@ public class IngressIT {
   }
 
   @Test
+  @ConditionalIgnore(condition = K8sVersionLessThan_1_14.class)
   public void get() {
     ingress = client.network().v1beta1().ingresses().inNamespace(session.getNamespace()).withName("ingress-get").get();
     assertThat(ingress).isNotNull();
   }
 
   @Test
+  @ConditionalIgnore(condition = K8sVersionLessThan_1_14.class)
   public void list() {
     IngressList aIngressList = client.network().v1beta1().ingresses().inNamespace(session.getNamespace()).list();
     assertNotNull(aIngressList);
@@ -71,6 +85,7 @@ public class IngressIT {
   }
 
   @Test
+  @ConditionalIgnore(condition = K8sVersionLessThan_1_14.class)
   public void update() {
     ingress = client.network().v1beta1().ingresses().inNamespace(session.getNamespace()).withName("ingress-update").edit()
       .editOrNewMetadata().addToAnnotations("foo", "bar").endMetadata().done();
@@ -80,11 +95,13 @@ public class IngressIT {
   }
 
   @Test
+  @ConditionalIgnore(condition = K8sVersionLessThan_1_14.class)
   public void delete() {
     assertTrue(client.network().v1beta1().ingresses().inNamespace(session.getNamespace()).withName("ingress-delete").delete());
   }
 
   @AfterClass
+  @ConditionalIgnore(condition = K8sVersionLessThan_1_14.class)
   public static void cleanup() {
     ClusterEntity.remove(IngressIT.class.getResourceAsStream("/ingress-it.yml"));
   }